+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### ✨ Added
 - handle nested for-loop
 
+### 🐛 Fixed
+- support windows OS
+
+
 ### 🔥 Removed
 - possibility to define multiple states for the same route